db persist

+++ backend/project_run.sh
@@ -5,7 +5,12 @@ echo "run in raw mode"
 mkdir share_data
 
 echo "=== recreate db ==="
-python3 merge.py || { echo "recreate db failed"; exit 1; }
+if [ -f "./9900-test-db.sqlite" ];then
+  echo "use shared_data db"
+  else
+  echo "use initialized db"
+  python3 merge.py || { echo "recreate db failed"; exit 1; }
+fi
 
 cp ./py_history/py_stock.py ./share_data/
 cp ./py_history/requirements.txt ./share_data/
